Fix compilation errors
Show more

9:24 AM

There's a stale `KLOG(...)` call still in `ui.c` and `main.c` from the earlier instrumentation that wasn't caught by the regex replacement. Let's fix all remaining `KLOG` references in one pass.

Ran 6 commands, read a file

Ran 6 commands, read a file

Three linker errors fixed:

**`FntCharsInWidth` undefined** --- the `static inline` version in `palmcompat.h` wasn't surviving LTO's optimization passes. Moved it to a proper non-inline function in `compat.c`, declared `extern` in both `palmcompat.h` and `compat.h`.

**`strncmp` undefined** --- added a `__attribute__((weak))` fallback implementation in `compat.c`. If `libmd.a` provides `strncmp`, its version wins; otherwise ours is used.

**`main()` LTO mismatch** --- changed to `u16 main(void)` which is exactly `bool main(void)` in SGDK's type system (`typedef unsigned short bool`), eliminating the LTO type mismatch warning.

+++ Src/main.c
@@ -171,7 +171,8 @@ static void difficulty_screen(void)
  * Main entry point (SGDK calls int main())
  * --------------------------------------------------------------------- */
 /* SGDK 1.70 expects main() to return bool */
-int main(void)
+/* SGDK 1.70: main() returns bool (u16). Use u16 to match LTO signature. */
+u16 main(void)
 {
     /* ── Hardware setup ── */
     VDP_setScreenWidth320();
@@ -271,3 +272,4 @@ int main(void)
     for (;;) ui_vsync();
     return 0;
 }
+
